Move tests over to testing project.
