+++ testing/src/test/scala/nl/tudelft/tydi_chisel/TydiStreamDriverTest.scala
@@ -3,7 +3,7 @@ package nl.tudelft.tydi_chisel
 import chisel3._
 import chisel3.experimental.BundleLiterals.AddBundleLiteralConstructor
 import chiseltest._
-import nl.tudelft.tydi_chisel.Conversions._
+import nl.tudelft.tydi_chisel_test.Conversions._
 import org.scalatest.flatspec.AnyFlatSpec
 
 class TydiStreamDriverTest extends AnyFlatSpec with ChiselScalatestTester {
