+++ testing/src/test/scala/nl/tudelft/tydi_chisel/chisel/BundleLiteralsSpec.scala
