+++ testing/src/test/scala/nl/tudelft/tydi_chisel/pipeline/PipelineExamplePlusTest.scala
@@ -4,7 +4,7 @@ import chisel3._
 import chisel3.experimental.BundleLiterals.AddBundleLiteralConstructor
 import chisel3.experimental.VecLiterals.{AddObjectLiteralConstructor, AddVecLiteralConstructor}
 import chiseltest._
-import nl.tudelft.tydi_chisel.Conversions._
+import nl.tudelft.tydi_chisel_test.Conversions._
 import nl.tudelft.tydi_chisel.{TydiProcessorTestWrapper, TydiTestWrapper}
 import org.scalatest.flatspec.AnyFlatSpec
